Some MSVC fixes. Further testing of multi threading needed, as application locks on shutting down (fails to deposit element for waiting queue. Seems boost.condition works in some strange way and does not unlock mutex. Tested with MSVC 2008 and boost 1.39

+++ Tests/Tutorial_1/main.cpp
@@ -26,7 +26,7 @@ public:
 	virtual int operator () ()
 	{
 		for(int k = 0; k<100;k++)
-			j = sin(j);
+			j = sin((float)j);
 		return core::taskmanager::Task::SECONDARY_THREAD;
 	}
 private:
